# openapi/fastbuyjson.yaml
openapi: 3.0.3
info:
  title: FastBuyJSON
  version: 1.0.0
  description: |
    JSON contract for LLM agents talking to e-commerce APIs.
    All commerce endpoints live under `/api/fastbuyjson`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/fastbuyjson
    description: Node.js demo server
  - url: http://localhost:8000/api/fastbuyjson
    description: Python demo server
...
security:
  - bearerAuth: []
  - {}
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/certificate:
    post:
      tags: [Auth]
      summary: Verify a client certificate
      operationId: verifyCertificate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [certificate]
              properties:
                certificate:
                  type: string
      responses:
        "200":
          description: Certificate session
          content:
            application/json:
              schema:
                type: object
                required: [session_id, expires_in]
                properties:
                  session_id:
                    type: string
                  expires_in:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    RefreshResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
...